Move curves package to framework/math
