+++ app/dance/movers/aggressive.go
@@ -3,7 +3,7 @@ package movers
 import (
 	"github.com/wieku/danser-go/app/beatmap/objects"
 	"github.com/wieku/danser-go/app/bmath"
-	"github.com/wieku/danser-go/app/bmath/curves"
+	"github.com/wieku/danser-go/framework/math/curves"
 	"github.com/wieku/danser-go/framework/math/vector"
 	"math"
 )
