+++ framework/math/curves/approximation.go
