+++ framework/math/curves/bezier.go
