+++ framework/math/curves/bezierapproximator.go
