+++ framework/math/curves/catmull.go
