+++ framework/math/curves/cirarc.go
