+++ framework/math/curves/curve.go
