+++ framework/math/curves/linear.go
