+++ framework/math/curves/multicurve.go
